refactor: Update tests.yml to fix timezone in dependabot.yaml
This commit updates the dependabot.yaml file to fix the timezone value in the schedule section. The previous value was set to "UTC", but it has been changed to "Etc/GMT" to accurately reflect the desired timezone. This change ensures that the schedule runs at the correct time according to the specified timezone.