feat: Add support for reconnecting in RPC disconnect method
This commit modifies the RPC disconnect method in the `echo.go` file to add support for reconnecting. It checks if the data parameter contains the string "reconnect" and disconnects the client with reconnection if true. Otherwise, it disconnects the client without reconnection. This enhancement improves the flexibility of the RPC disconnect functionality.