feat: Update timeouts in smoke_test.dart to improve RPC testing
This commit updates the timeouts in the smoke_test.dart file to improve RPC testing. The previous timeouts were [1000, 100, 10, 100, 0], and they have been changed to [200, 50, 10, 25, 0]. This change ensures that the RPC requests are sent and received within the specified time intervals, enhancing the reliability and performance of the RPC functionality.