refactor: Update WebSocket handling and Protobuf transport in smoke_test.dart
This commit refactors the WebSocket handling and adds support for Protobuf transport in the smoke_test.dart file. It improves the RPC testing by updating the timeouts from [1000, 100, 10, 100, 0] to [200, 50, 10, 25, 0]. Additionally, it introduces reconnecting functionality in the RPC disconnect method in the echo.go file. The client is now disconnected with reconnection if the data parameter contains the string "reconnect". This enhancement improves the flexibility and reliability of the RPC functionality.