refactor: Update BenchmarkController to use abstract base class
This commit updates the `BenchmarkController` class to use an abstract base class called `BenchmarkControllerBase`. This change allows for better code organization and separation of concerns. The `BenchmarkControllerBase` class contains common properties and methods that are shared by different benchmark implementations. By using an abstract base class, it becomes easier to add new benchmark implementations in the future. This refactor improves the maintainability and extensibility of the codebase.